Split off Blob and Data.
